Added .gitignore and restructured dir
To help with deployment, GitHub actions, and other JS/TS stuff, I have restructured the directory so the frontend is at the root of the directory.  This allows scripts to easily locate the files to build, also allows git-ignore to work correctly and hide IDE, node_modules and other items like that.